+++ jambeth/jambeth-service/src/main/java/de/osthus/ambeth/merge/model/IEntityMetaData.java
@@ -16,7 +16,7 @@ public interface IEntityMetaData
 
 	PrimitiveMember getIdMember();
 
-	PrimitiveMember getIdMemberByIdIndex(byte idIndex);
+	PrimitiveMember getIdMemberByIdIndex(int idIndex);
 
 	byte getIdIndexByMemberName(String memberName);
